use absolute imports
Start to pick apart the stacks.js to demo the absolute imports
