+++ webpack.config.js
@@ -1,21 +1,31 @@
 const HtmlWebPackPlugin = require("html-webpack-plugin");
+const path = require('path');
 
 module.exports = {
   entry: {
     main: "./src/stacks.js"
   },
   devtool: "source-map",
+  resolve: {
+    modules: [path.resolve(__dirname, './src'), 'node_modules'],
+    extensions: ['.js', '.jsx'],
+    alias: {
+      "*": path.resolve(__dirname, './src')
+    }
+  },
   module: {
     rules: [
       {
-        test: /\.js$/,
+        test: [/\.js$/, /\.jsx$/],
         exclude: /node_modules/,
-        use: {
-          loader: "babel-loader"
-        }
+        use: [
+          {
+            loader: "babel-loader"
+          }
+        ]
       },
       {
-        test: /\.html$/,
+        test: [/\.html$/],
         use: [
           {
             loader: "html-loader"
